// src/app/store/action-types/people-action.types.ts
export enum PeopleActionType {
    PEOPLE_LIST_LOAD = '[People] Load People',
    PEOPLE_LIST_LOAD_SUCCESS = '[People] Load People Success',
    PEOPLE_LIST_LOAD_FAILURE = '[People] Load People Failure',

    CONVERSATIONS_LIST_LOAD = '[People] Load Conversations',
    CONVERSATIONS_LIST_LOAD_SUCCESS = '[People] Load Conversations Success',
    CONVERSATIONS_LIST_LOAD_FAILURE = '[People] Load Conversations Failure',

    CONVERSATION_CREATE = '[People] Create Conversation',
    CONVERSATION_CREATE_SUCCESS = '[People] Create Conversation Success',
    CONVERSATION_CREATE_FAILURE = '[People] Create Conversation Failure',

    PEOPLE_LIST_LOADING_FALSE = '[People] People List Loading False',
}